% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\name{calls}
\alias{calls}
\alias{call_time}
\alias{call_start}
\alias{call_end}
\alias{call_duration}
\alias{call_timings}
\title{Function call properties}
\usage{
call_time(.time, .form = "\%Y-\%m-\%d \%H:\%M:\%S")

call_start(.fun = as.character(sys.call(-1L)), .start = Sys.time())

call_end(.fun = as.character(sys.call(-1L)), .start, .end = Sys.time())

call_duration(.start, .end, ...)

call_timings(.start, .end = Sys.time(), ...)
}
\arguments{
\item{.time, .form, .fun, .start, .end, ...}{Arguments.}
}
\description{
These functions collate information related to function calls.
}
\author{
Edward Lavender
}
