% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cats-iter.R
\name{cats_iter}
\alias{cats_iter}
\alias{cat_time}
\alias{cat_line}
\alias{cat_row}
\alias{cat_next}
\alias{cat_do}
\title{\code{cat} functions for iterative workflows}
\usage{
cat_time()

cat_line()

cat_row(.index)

cat_next(.index, .verbose)

cat_do(..., .verbose)
}
\arguments{
\item{.index, .verbose, ...}{Arguments.}
}
\description{
\code{cat} functions for iterative workflows
}
\details{
These functions were motivated by the \code{patter.workflows} package.
}
