% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{lists}
\alias{lists}
\alias{list_compact}
\alias{list_merge}
\alias{list_args}
\title{\code{list} functions}
\usage{
list_compact(l)

list_merge(...)

list_args(default = list(), dots = list(), user = list())
}
\arguments{
\item{l, ..., default, dots, user}{Arguments.}
}
\description{
\code{list} functions
}
\author{
Edward Lavender
}
