% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.build-package.R
\docType{package}
\name{proj.build}
\alias{proj.build-package}
\alias{proj.build}
\title{\code{\link{proj.build}}}
\description{
Scientific project development and management.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/edwardlavender/proj.build/}
  \item Report bugs at \url{https://github.com/edwardlavender/proj.builds/issues/}
}

}
\author{
\strong{Maintainer}: Edward Lavender \email{edward.lavender@eawag.ch} (\href{https://orcid.org/0000-0002-8040-7489}{ORCID})

}
