% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinks.R
\name{sinks}
\alias{sinks}
\alias{sink_open}
\alias{sink_close}
\title{\code{sink()} functions}
\usage{
sink_open(log.txt = NULL)

sink_close(log.txt = NULL)
}
\arguments{
\item{log.txt}{(optional) A \code{character} string that specifies the path to a \code{log.txt} file.}
}
\description{
\code{sink()} functions
}
